build(coscrad-frontend): migrate cypress to 10 (#191)
* build(coscrad-frontend): migrate cypress to 10

* test(coscrad-frontend): remove skipped test
