--- apps/coscrad-frontend-e2e/cypress.json
@@ -1,13 +1,0 @@
-{
-  "fileServerFolder": ".",
-  "fixturesFolder": "./src/fixtures",
-  "integrationFolder": "./src/integration",
-  "modifyObstructiveCode": false,
-  "supportFile": "./src/support/index.ts",
-  "pluginsFile": false,
-  "video": true,
-  "videosFolder": "../../dist/cypress/apps/coscrad-frontend-e2e/videos",
-  "screenshotsFolder": "../../dist/cypress/apps/coscrad-frontend-e2e/screenshots",
-  "chromeWebSecurity": false,
-  "baseUrl": "http://localhost:4200"
-}
