+++ apps/coscrad-frontend-e2e/src/e2e/app.cy.ts
@@ -1,0 +1,7 @@
+describe('coscrad-frontend', () => {
+    beforeEach(() => cy.visit('/'));
+
+    it('should display welcome message', () => {
+        cy.contains('coscrad', { matchCase: false });
+    });
+});
