--- apps/coscrad-frontend-e2e/src/integration/app.spec.ts
@@ -1,8 +1,0 @@
-
-describe('coscrad-frontend', () => {
-  beforeEach(() => cy.visit('/'));
-
-  it('should display welcome message', () => {
-    cy.contains('coscrad',{matchCase:false})
-  });
-});
