+++ apps/tng-dictionary-e2e/cypress.config.ts
@@ -1,0 +1,19 @@
+import { defineConfig } from 'cypress';
+import { nxE2EPreset } from '@nrwl/cypress/plugins/cypress-preset';
+
+const cypressJsonConfig = {
+    fileServerFolder: '.',
+    fixturesFolder: './src/fixtures',
+    video: true,
+    videosFolder: '../../dist/cypress/apps/tng-dictionary-e2e/videos',
+    screenshotsFolder: '../../dist/cypress/apps/tng-dictionary-e2e/screenshots',
+    chromeWebSecurity: false,
+    specPattern: 'src/e2e/**/*.cy.{js,jsx,ts,tsx}',
+    supportFile: 'src/support/e2e.ts',
+};
+export default defineConfig({
+    e2e: {
+        ...nxE2EPreset(__dirname),
+        ...cypressJsonConfig,
+    },
+});
