+++ apps/tng-dictionary-e2e/src/support/e2e.ts
