--- apps/tng-radio-hub-e2e/cypress.json
@@ -1,12 +1,0 @@
-{
-    "fileServerFolder": ".",
-    "fixturesFolder": "./src/fixtures",
-    "integrationFolder": "./src/integration",
-    "modifyObstructiveCode": false,
-    "supportFile": "./src/support/index.ts",
-    "pluginsFile": false,
-    "video": true,
-    "videosFolder": "../../dist/cypress/apps/tng-radio-hub-e2e/videos",
-    "screenshotsFolder": "../../dist/cypress/apps/tng-radio-hub-e2e/screenshots",
-    "chromeWebSecurity": false
-}
