+++ apps/tng-radio-hub-e2e/tsconfig.json
@@ -6,5 +6,5 @@
         "allowJs": true,
         "types": ["cypress", "node"]
     },
-    "include": ["src/**/*.ts", "src/**/*.js"]
+    "include": ["src/**/*.ts", "src/**/*.js", "cypress.config.ts"]
 }
