--- cypress.json
@@ -1,4 +1,0 @@
-{
-    "baseUrl": "http://localhost:4200/",
-    "integrationFolder": "cypress/e2e"
-}
